Man what a mess.  My excuse will be that sometimes things have to get worse to get better :)
In this case, I felt like the dumper/Writer was doing too much.  It was
doing object/bean traversal and serialization.  This commit explored
actually pulling out the traversal concern into a Walker.  SO,
the Walker is responsible for knowing how to walk over an object
graph, and it takes a WalkerVisitor which has various callback
methods on it to be called during object traversal (very SAX-y).
Then, there is a JsonProcessor (implements WalkerVisitor...will
rename to something better later) that handles the actual JSON
specific serialization logic.

I didn't want to drastically break what was there, so the Writer
was largely left alone, except to add a parallel toJson method that
took an actual Writer (as opposed to the existing one that took
a JsonWriter) which was a quick copy-paste for the most part of the
other method (remember I did say it was ugly :)).  I wanted to get
to a quick Proof of Concept before really ripping out the guts
and wanted to be able to run the tests against both approaches as
I went.  In the end the Walker approach is about 13-15% faster than
the original Writer approach....and it has the concerns split out.

I considered not checking this in yet (it even has a failing test
on the Json stuff b/c it doesn't handle circular refs yet), but
decided it needed to go in...then I can start really ripping things
apart.  I suppose i could create a branch, but...ah well.