Tweaked the JsonSerializerTest to focus squarly on performance. Added memory printouts (need to move them to asserts or an average calc).  Added net.sf.json and a test for it for comparison.
So at the end of the day, performance looks ok.  I do need to split the perf tests into
their own fixture so I can un-comment the other tests.

Our worst case was using the "WithPath" Policy and it came in at
18ms and 975KB (burst) memory usage.
net.sf.json came it at
150ms and 1470KB (burst) memory usage (just for the JSON object anyway).
Our best case was the "Do Nothing" TrackingPolicy and it came in at:
11.5ms and 521KB (burst) memory usage.

Bunch of cleanup still to do, but at least is is on track perf-wise and I feel
OK to drop the other code line.