Nuked the JsonWriter and all interface methods referencing it.  Showed some interesting things, like the Collection and NVPair Serializers don't do anything in the new walker approach.
Will focus on getting the json visitor side of things straightened out for now, then
move on to continuing with the walker refactoring.  The walker may wind up having a similar
registry notion for looking up the right strategry based on type.  OR it might just do
an adapter/translator registry to create the right kind of iterable for the more complex
types:  either an NVPairs iterable for beans, or a straight object iterable for collections.